--- src/Developer.java
@@ -1,29 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package task06;
-
-/**
- *
- * @author main
- */
-public class Developer {
-    
-    private String qualification;
-    private int salary;
-    private int productivity;
-
-    public String getQualification() {
-        return qualification;
-    }
-
-    public int getSalary() {
-        return salary;
-    }
-        
-    public int getProductivity() {
-        return productivity;
-    }
-}
